Improved PlayerManager script
In the Awake function I've put in an extra check to see if the collected count of Aku Aku's is equal to three, meaning we left the previous level while invinsible. So we need to withdraw one Aku Aku from the count so once the player collects another Aku AKu he can go back into the invincibility state.