Improved default crate bounce
Added an rigidbody to the default crate and check it's Y velocity to see if it's higher then three and if it's allowed to bounce up. If so I give the crate extra upwards momentum to make it seem it bounced off the nitro crate a bit . Once it reaches it's velocity back to zero again we allow it to bounce up again.