Nextlevel detecting
Build in a raycast check to see if the player hits another level and route in the direction he's trying to walk. If it hits something we display it in the debug loggin.

TO DO:
Find a way to move the player to the location it found with the raycast based on the players direction input.