Potentional Camera script
Followed Brackeys tutorial on how to keep multiple players within the field of view. I've used this to keep my player in field of view of the camera and zoom in once the player moves and out when standing still.

To do:
Fix snapping effect when camera goes from zooming out to in and vise versa